+++ lib/view/page/comment_form/comment_form_page.dart
@@ -4,9 +4,9 @@ import 'package:iris_flutter/config/config.dart';
 import 'package:iris_flutter/view/comm/custom_appbar.dart';
 import 'package:iris_flutter/view/comm/form/basic_form.dart';
 import 'package:iris_flutter/view/comm/form/text_form.dart';
-import 'package:iris_flutter/view/comm/register_button.dart';
+import 'package:iris_flutter/view/comm/form/register_button.dart';
 import 'package:iris_flutter/view/controller/comment_form/comment_form_controller.dart';
-import 'package:iris_flutter/view/page/Info_form/image_form_carousel.dart';
+import 'package:iris_flutter/view/comm/form/image_carousel_form.dart';
 
 class CommentFormPage extends StatefulWidget {
   const CommentFormPage({Key? key}) : super(key: key);
@@ -69,7 +69,7 @@ class _CommentFormState extends State<CommentFormPage> {
               ),
               const Padding(padding: EdgeInsets.only(bottom: 15)),
               // 제보 사진 추가
-              ImageFormCarousel(
+              ImageCarouselForm(
                 title: '제보',
                 controller: commentController,
               ),
